backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
anthropic==0.18.1
openai==1.12.0
firebase-admin==6.4.0
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
python-dotenv==1.0.0
Pillow==10.2.0
PyPDF2==3.0.1
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0
sentry-sdk==1.40.0
prometheus-client==0.19.0
